pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["MarsBroshok96 <marsbroshok1996@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.1.7"
python-dotenv = "^1.0.0"
dj-database-url = "^1.2.0"
gunicorn = "^20.1.0"
django-bootstrap4 = "^22.3"
whitenoise = "^6.4.0"
psycopg2-binary = "^2.9.5"
factory-boy = "^3.2.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
coverage = "^7.2.1"
django-extensions = "^3.2.1"
ipython = "^8.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
